backend/requirements.txt
<<<<<<< HEAD
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
openpyxl==3.1.2
numpy==1.26.4
transformers==4.40.2
torch>=2.0.0
pandas
python-docx
PyMuPDF
chromadb
=======
uvicorn
python-dotenv
requests
openai
pymupdf         # For PDF text extraction
pillow          # For image handling
pandas          # For Excel parsing
openpyxl        # Excel engine
python-multipart
google-generativeai
>>>>>>> origin/Sparsh
